Move tests to match refactored DGF files
In https://github.com/communitiesuk/funding-service/pull/556 I moved the code,
but neglected to move the tests as well. This does a bit of a tidy up so that
test paths match the main app code paths.
